Lots of changes
The system scripts are now up to date with the interpreter, T.A00.052 (from the PQ3 demo).
Of course, that was the point where timer bugs became common in SCI. I took those into account in GAME.SC and rSpeedTest.sc.

Also, Ego will now be played by Sonny Bonds, replacing Roger Wilco in both appearance and functionality. It'd be nice to have an original ego view, but graphics are not my thing.