Identified more procedures, replaced Studio defines with Sierra ones
More globals and procedures have been identified. Also, a lot of unused SQ4 globals have been removed, so a LOT of scripts have to be updated to reflect that.

A major improvement is that PalInit's globals have been fully identified, as has the procedure that checks if the game is EGA or VGA.
I may think about adding EGA support soon...