Updated interpreter to T.A00.178
The template's got a major upgrade: a newer interpreter!

As a result, the system scripts have been updated to support this, and are a bjt closer to the SCI16 source.

The reason for upgrading the interpreter is to prepare for future EGA support. Using PQ3's interpreter makes the most sense, since it has both EGA and VGA versions.

One important thing: the updated interpreter relies on global 101 to be set to TRUE, otherwise it will crash with Error 11 shortly after startup! It seems to be a dongle-based copy protection.