Template overhauled!
This template has been completely overhauled. It is now built on top of the Mixed-Up Fairy Tales demo.
As a result, the interpreter is newer (dated December 1991, if the timestamp is any indication), and has separate music and digital audio drivers. That makes the GMBLAST driver no longer necessary.

Ego is now played by the white boy from MUFT.